Refactor handling of trailing dots in filenames
Though the previous code did have a workaround for a trailing dot in
the filenme, there were still issues.

The main issue was that, if the backing filesystem was a FAT fs (FAT32
or EXFAT, doesn't matter), or say a mounted shared fs, when the
HCB_PREFIX (originally .pxovl. prefix) is created, the trailing dot
would be eaten.

Though the HCB_PREFIX1 workaround was put in to detect the file
without the trailing dot, rmdir would not work. As an attempt to
unlink the file that had it's trailing dot eaten.

Ultimately, being seemingly arbitrarily set, I removed the HCB_PREFIX1
attempt to workaround trailing dots, and changed HCB_PREFIX to
.pxovl._ so that the underscore is at least the trailing character in
the entry for the directory it resides in (I'll leave aside the
possibility that backwards compatibility tied it in. Obviously, this
change breaks backwards compatibility).

I then modified xfrm_to_disk so that it detects trailing dots on each
filename component of the passed path (so in between slashes / if
necessary), and handle/escape them like the special characters.

A minor change was also to alter posixovl_getxattr to be consistent in
targetting the path with the escaped filename components.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Changes to be committed:
#	modified:   posixovl.c
#