Fix symlink handling, not requiring normalized path for target & xattr
The symlink code tries to normalise the symlink for the backing
filesystem. But this is unnecessary, as a) symlinks can be "dangling",
where the path is not reachable. Having this so on the backing
filesystem is fine b) accessing the symlink on the mounted paths, if
resolves to another path on the mounted path, will access the right
path through the posixovl filesystem.

Also, fixed the remaining xattr relevant operations for the source
path.